backend/requirements.txt
fastapi
uvicorn
pydantic
httpx
python-dotenv
sqlalchemy
aiosqlite
passlib[bcrypt]
python-jose[cryptography]
python-multipart
